Corrected progressBar
