+++ .gitignore
@@ -22,4 +22,3 @@ yarn-debug.log*
 yarn-error.log*
 
 .docz
-package-lock.json
