+++ .npmrc
@@ -1,3 +1,2 @@
 save-exact=true
 loglevel=warn
-package-lock=false
