+++ src/ignitus-StudentSignUp/Components/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as Signup } from '../profile/Components/profile';
+export {default as Signup} from '../signUp/Components/signUp';
