+++ src/ignitus-StudentSignUp/email/Styles/style.scss
@@ -1,6 +1,6 @@
 // import ignitus token variables
 @import './src/shared/styles/ignitus-components-common.scss';
-@import '../../profile/Styles/style.scss';
+@import '../../signUp/Styles/style.scss';
 
 .email {
   height: 48rem;
